% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_dist.R
\name{aa_dist}
\alias{aa_dist}
\title{Amino acid distance}
\usage{
aa_dist(a = NA, b = NA, dists = "grantham", extra.info = F)
}
\arguments{
\item{a}{The first sequence}

\item{b}{The second sequence}

\item{dists}{Distance metric to be used (defaults to 'grantham')}

\item{extra.info}{Option to return more information as a data frame}
}
\value{
Grantham score summed across aa differences in sequences
}
\description{
Given two aligned amino acid sequences, calculate the Grantham distance.
}
\examples{
aa_dist(c('AAT', 'VPG'), extra.info = T)
}
