% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_info.R
\name{aa_info}
\alias{aa_info}
\title{Amino acid information}
\usage{
aa_info()
}
\value{
data frame with amino acid information
}
\description{
Returnes a data frame with information on each amino acid including
a biophysics-based colouring scheme
}
\examples{
aa_info()
}
