% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_node_dates.R
\name{add_node_dates}
\alias{add_node_dates}
\title{Add node dates}
\usage{
add_node_dates(tree_dat = NA, mrsd = NA, pst = FALSE)
}
\arguments{
\item{tree_dat}{Data frame describing tree structure}

\item{mrsd}{date in format "YYYY-MM-DD"}

\item{pst}{logical, does dataframe represent posterior sample of trees}
}
\value{
version of tree_dat with columns `date` and `date_format`
}
\description{
Calculate dates of internal nodes of tree given tree structure and the
most recent sampling date (mrsd). Can be done for single tree or for a
posterior sample of trees (pst).
}
\details{
`tree_dat` should have column `x` and in the case of PST, `.id` too.
}
