% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asn_seq_cluster.R
\name{asn_seq_cluster}
\alias{asn_seq_cluster}
\title{Assign sequence cluster}
\usage{
asn_seq_cluster(seq_vector = NA, threshold = 0.02, method = "single")
}
\arguments{
\item{seq_vector}{Character vector representing aligned nucleotide sequences}

\item{threshold}{threshold similarity to pass to bioseq::seq_cluster()}

\item{method}{clustering method to pass to bioseq::seq_cluster()}
}
\description{
Assign clusters on the basis of sequence similarity using the seq_cluster() function from
the bioseq package. Re-label
}
