% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_index.R
\name{association_index}
\alias{association_index}
\title{Association index}
\usage{
association_index(tree_dat = NA, trait_var = "trait", add_col = FALSE)
}
\arguments{
\item{tree_dat}{Data frame describing tree structure}

\item{trait_var}{Name of column with discrete trait}

\item{add_col}{Logical - whether to return tree_dat with added column showing assocaition metric per node}
}
\value{
Association metric
}
\description{
Calculate the association index (Wang et al. 2001 doi: 10.1128/JVI.75.23.11686-11699.2001)
}
\details{
The function requires a data frame with column 'node' and 'parent'.
Such a data frame is generated using ggtree::fortify
Further requires a column with a discrete trait in which to explore phylogenetic signal
}
