% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avian_orders.R
\name{avian_orders}
\alias{avian_orders}
\title{Avian orders}
\usage{
avian_orders(input = NA)
}
\arguments{
\item{input}{by default, all info will be returned though a specific order can be supplied to minimise info returned}
}
\description{
Information to remind on birds comprising avian orders.
}
