% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beast_summarise_rates_bssvs.R
\name{beast_summarise_rates_bssvs}
\alias{beast_summarise_rates_bssvs}
\title{Summarise BEAST BSSVS rates}
\usage{
beast_summarise_rates_bssvs(
  rates_dat = NA,
  mean_rate = 1,
  threshold_cond = 0.5,
  interval = "hpd"
)
}
\arguments{
\item{rates_dat}{Data frame version of log file}

\item{mean_rate}{Mean overall rate for trait transitions across tree}

\item{threshold_cond}{Value above which to retain conditional means. Defaults to 0.5.}

\item{interval}{'hpd' (default) for highest posterior density interval or
'credible' for equal-tailed credible interval interval.}
}
\value{
summary dataframe for transitions and NA matchining dataframe for non-transitions
}
\description{
Summarise info in BEAST generated log file from a run
with discrete traits and a BSSVS network.
}
