% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpg_count.R
\name{cpg_count}
\alias{cpg_count}
\title{CpG count}
\usage{
cpg_count(sequence = NA)
}
\arguments{
\item{sequence}{String representation of nucleotide sequence}
}
\description{
Count CpG in a nucleotide sequence.
}
\examples{
sequence <- 'CATGCTGTGGTCGCGTAGGGTTATTCGTCGTAATTGCAGCT'
cpg_count(sequence, method = 2, res_expand = T)
}
