% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_branch_vars.R
\name{generate_branch_vars}
\alias{generate_branch_vars}
\title{Generate branch variables}
\usage{
generate_branch_vars(tree_dat = NA, variable_type = "identity", prefix = NA)
}
\arguments{
\item{tree_dat}{Data frame describing tree structure}

\item{variable_type}{'identity' for whether branch occurs in evolutionary history. 'difference' for whether branch separated taxa or not}

\item{prefix}{prefix to attach to}
}
\value{

}
\description{
Create columns of logical variables that indicate whether branches
of a phylogenetic tree occur in the evolutionary history of a taxon
(variable_type = 'identity') or separate two taxa (variable type =
'difference').
}
\details{
Function calls functions node_ancestors() and nodes_relationship()
}
