% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grantham_dists.R
\name{grantham_dists}
\alias{grantham_dists}
\title{Grantham distances}
\usage{
grantham_dists()
}
\value{
dataframe including Grantham distances
}
\description{
Function provides data for other functions to use on Grantham distances
https://en.wikipedia.org/wiki/Amino_acid_replacement
}
\examples{
grantham_dists()
}
