% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_string_diff.R
\name{list_string_diff}
\alias{list_string_diff}
\title{List string differences}
\usage{
list_string_diff(
  a = "ATTCGA-",
  b = "attTGTT",
  exclude = c("-", "?"),
  ignore.case = TRUE
)
}
\arguments{
\item{a}{string a}

\item{b}{string b}

\item{exclude}{vector of characters to ignore when identifying differences}

\item{ignore.case}{logical, whether or not to ignore case in comparison}
}
\value{
list of differences between two strings
}
\description{
List string differences
}
