% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_branching_index.R
\name{local_branching_index}
\alias{local_branching_index}
\title{Local branching index}
\usage{
local_branching_index(tree_dat = NA, scale = NA)
}
\arguments{
\item{tree_dat}{Data frame describing tree structure}

\item{scale}{Scaling parameter}
}
\description{
Calculate the local branching index (Neher et al. 2014 eLife doi: 10.7554/eLife.03568
For each node in a phylogeny, the LBI is the exponentially weighted tree length
surrounding a node. Higher LBI indicated rapid branching and clade expansion.
}
\details{
The function requires a data frame with column 'node' and 'parent'.
Such a dataframe is generated using ggtree::fortify
}
