% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_branching_index_set.R
\name{local_branching_index_set}
\alias{local_branching_index_set}
\title{Local branching index based on subset of tips}
\usage{
local_branching_index_set(
  tree_dat = NA,
  tree = NA,
  taxa = NA,
  nodes = NA,
  use_nodes = F,
  scale = NA
)
}
\arguments{
\item{tree_dat}{Data frame describing tree structure}

\item{tree}{Phylo class object for tree associated with tree_dat}

\item{taxa}{Vector of taxon labels forming subset}

\item{nodes}{Optionally subset can be supplied as vector of terminal nodes}

\item{use_nodes}{Logical should nodes be used instead of taxa labels}

\item{scale}{Scaling parameter to be passed to local_branching_index function}
}
\value{
tree_dat with additional column for lbi
}
\description{
Calculate the local branching index (LBI, Neher et al. 2014 eLife doi: 10.7554/eLife.03568).
This version of the function opperates by calculating the LBI using a user supplied subset
of all taxa/tips/terminal nodes. Branches leading towards tips excluded from this set are
ignored, as are branching events that only exist due to the samping of taxa outside the
set.
}
\details{
For each node in a phylogeny, the LBI is the exponentially weighted tree length
surrounding a node. Higher LBI indicated rapid branching and clade expansion.

The function requires a data frame with column 'node' and 'parent'.
Such a dataframe is generated using ggtree::fortify
}
