% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_alignments.R
\name{merge_alignments}
\alias{merge_alignments}
\title{Merge alignments
With a list of fasta files, function reads in each one, runs the function
process_alignment(). process_alignment() will trim sequences and remove
insertions as well as tidying names/sequences. If the length of tidied
sequences match, function will then rbind() alignments together.}
\usage{
merge_alignments(
  alignment_files = files_pb2,
  rm.duplicates = TRUE,
  isolate.id.column = TRUE,
  suffix = NA
)
}
\arguments{
\item{alignment_files}{list of fasta files}

\item{rm.duplicates}{passed to process_alignment()}

\item{isolate.id.column}{passed to process_alignment()}

\item{suffix}{passed to process_alignment()}
}
\value{
dataframe with tidied genetic data
}
\description{
Merge alignments
With a list of fasta files, function reads in each one, runs the function
process_alignment(). process_alignment() will trim sequences and remove
insertions as well as tidying names/sequences. If the length of tidied
sequences match, function will then rbind() alignments together.
}
