% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nglyc_locate.R
\name{Nglyc_locate}
\alias{Nglyc_locate}
\title{N-linked glycosylation site locations}
\usage{
Nglyc_locate(sequence, allow_dash = F, allow_X = F)
}
\arguments{
\item{sequence}{string representing amino acid sequence}

\item{allow_dash}{logical, should dash be allowed in the middle position of the motif}

\item{allow_X}{logical, should X be allowed in the middle position of the motif}
}
\value{
the count of N-linked glycosylation sites identified
}
\description{
Given an amino acid sequence, identify the positions of potential N-linked glycosylation sites.
The N-linked glycosylation motif is N-X-S/T or Asn-X-Ser/T where X is any amino acid
except Proline (Pro/P).
}
\examples{
sequence <- 'MDTICIGHAYNNPALEKNKSSTDATNLTYYG'
Nglyc_locate(sequence)
}
