% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_ancestors.R
\name{node_ancestors}
\alias{node_ancestors}
\title{Node ancestors}
\usage{
node_ancestors(tree_dat = NA, node_id = NA, keep_target_node = FALSE)
}
\arguments{
\item{tree_dat}{Data frame describing tree structure}

\item{node_id}{The node from which to ID all ancestors}
}
\value{
Vector of ancestral nodes
}
\description{
For any node in rooted phylogeny, ID all ancestor nodes back to the root.
The function requires a data frame with column 'node' and 'parent'.
Such a dataframe is generated using ggtree::fortify
}
