% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_descendants.R
\name{node_descendants}
\alias{node_descendants}
\title{Node descendants}
\usage{
node_descendants(tree_dat = NA, node_id = NA, tips_only = FALSE)
}
\arguments{
\item{tree_dat}{Data frame describing tree structure}

\item{node_id}{The node from which to ID all descendants}

\item{tips_only}{Logical - return only tip nodes instead of all descendants - FALSE by default}
}
\value{
Vector of descendant nodes
}
\description{
For an internal node in rooted phylogeny, ID all descendant nodes.
The function requires a data frame with column 'node' and 'parent'.
Such a dataframe is generated using ggtree::fortify
}
