% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes_relationship.R
\name{nodes_relationship}
\alias{nodes_relationship}
\title{Nodes relationship}
\usage{
nodes_relationship(
  tree_dat = NA,
  taxa = NA,
  nodes = NA,
  use_nodes = F,
  relationship = "path"
)
}
\arguments{
\item{tree_dat}{Data frame describing tree structure}

\item{taxa}{The taxa from which to explore relationship}

\item{nodes}{The nodes from which to explore relationship (if use_nodes != TRUE)}

\item{use_nodes}{Select TRUE to use node numbers rather than taxa labels}

\item{relationship}{The type of relationship to return ('path', 'mrca' or 'shared' )}
}
\value{
Nodes reflecting an aspect of evolutionary relationship
}
\description{
For any two nodes in rooted bifurcating phylogeny, get ancestral nodes and
return an output based on this: vector of nodes forming path through tree
between taxa, node representing their most recent common ancestor, or the
vector of nodes representing their shared evolutionary history.
The function requires a data frame with column 'node' and 'parent'.
Such a data frame is generated using ggtree::fortify
}
