% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_avian_species.R
\name{parse_avian_species}
\alias{parse_avian_species}
\title{Parse avian species}
\usage{
parse_avian_species(dat = NA, species = "species")
}
\arguments{
\item{dat}{dataframe with column}

\item{species_col}{the name of the column with location info to be parsed}
}
\value{
dataframe with tidied species
}
\description{
Tidy up some commonly occurring errors in bird species names in influenza
sequence data bases and match to an avian order when possible
}
