% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_locations.R
\name{parse_locations}
\alias{parse_locations}
\title{Parse locations}
\usage{
parse_locations(dat = NA, loc_col = "loc_code", region_type = "regions_WHO")
}
\arguments{
\item{dat}{dataframe containing a column which includes location info to parse}

\item{loc_col}{the name of the column with location info to be parsed}

\item{region_type}{Approach to group nation states into larger global regions}
}
\value{
dataframe with columns representing parsed location and wider region
}
\description{
Parse locations
}
