% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsimony_score.R
\name{parsimony_score}
\alias{parsimony_score}
\title{Parsimony score}
\usage{
parsimony_score(
  tree_dat = NA,
  trait_var = "trait",
  prefer_intersect = TRUE,
  return_df = FALSE
)
}
\arguments{
\item{tree_dat}{Data frame describing tree structure}

\item{trait_var}{Name of column with discrete trait}

\item{prefer_intersect}{Logical to pass to tips_2_root function}

\item{return_df}{Logical - should dataframe with new columns be returned}
}
\description{
Calculate the parsimony score for a discrete trait labelled in tips. This is the
minimum number of trait changes required to give the pattern observed in the tips
and is a measure of phylogeny and trait correlation. Lower parsimony scores
indicate higher phylogeny-trait correlation.
}
