% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_clusters.R
\name{phylo_clusters}
\alias{phylo_clusters}
\title{Phylogenetic clusters}
\usage{
phylo_clusters(
  tree_dat,
  t = NA,
  w = 1,
  suppress_singletons = TRUE,
  suppress_doubles = FALSE
)
}
\arguments{
\item{tree_dat}{Data frame describing tree structure}

\item{t}{Threshold value for branch lengths defining clusters}

\item{suppress_singletons}{Logical - whether or not to require greater isolation of singletons}
}
\value{
Input data frame with addition of column with clusters
}
\description{
Given a data frame describing tree structure and branch lengths and a
defined threshold, determine clusters on phylogeny.
}
