% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_alignment.R
\name{process_alignment}
\alias{process_alignment}
\title{Process alignment}
\usage{
process_alignment(
  seq_dat = NA,
  rm.duplicates = TRUE,
  isolate.id.column = FALSE,
  suffix = NA,
  seq_trim.start = TRUE,
  seq_trim.end = TRUE,
  seq_insertion.threshold = NULL
)
}
\arguments{
\item{seq_dat}{data frame with columns seq.name and seq.text}

\item{rm.duplicates}{option to remove duplicates of sequence labels}

\item{isolate.id.column}{option to return an isolate ID column}

\item{seq_trim.start}{passed to process_seq() - option to trim before first start codon}

\item{seq_trim.end}{passed to process_seq() - option to trim beyond last stop codon}

\item{seq_insertion.threshold}{passed to process_seq() - proportion of bases with '-' required to cut position}
}
\description{
Function performs some standardisation of names, sequences . Function requires as input a
data frame with sequence names in column 'seq.name and nucleotides sequences
}
