% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_meta.R
\name{process_meta}
\alias{process_meta}
\title{Process metadata}
\usage{
process_meta(
  meta_dat = NA,
  rm.duplicates = TRUE,
  drop.laboratory = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{meta_dat}{data frame with columns 'Isolate_Id', 'Isolate_Name', 'Subtype', Location','Host', 'Collection_Date', 'Domestic_Status'.}

\item{rm.duplicates}{option to drop duplicate isolate IDs}

\item{drop.laboratory}{option to drop any isolates labelled as laboratory or recombinant from output. Defaults to TRUE}

\item{verbose}{option to print information on variables as they are created. Defaults to FALSE}
}
\description{
Process metadata
}
