% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_seq.R
\name{process_seq}
\alias{process_seq}
\title{Process sequences}
\usage{
process_seq(
  seqs = NA,
  trim.start = TRUE,
  trim.end = TRUE,
  insertion.threshold = NULL
)
}
\arguments{
\item{seqs}{vector of aligned sequences in character strings}

\item{trim.start}{option to trim before first start codon}

\item{trim.end}{option to trim beyond last stop codon}

\item{insertion.threshold}{max. proportion of bases with '-' that can survive insertion strip}
}
\description{
Process an vector containing aligned sequences as character strings. Function will remove
non-standard bases, optionally trim to protein coding regions, and optionally will remove
possible insertions. A threshold can be provided with which to remove insertions. For example
if
}
