% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range01.R
\name{range01}
\alias{range01}
\title{Re-scale to range 0-1}
\usage{
range01(x = NA, ...)
}
\arguments{
\item{x}{Data frame describing tree structure}
}
\value{
Re-scaled numeric vector
}
\description{
Re-scale a numeric vector such that the minimum number is 0 and maximum is 1.
}
