% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root_2_tips.R
\name{root_2_tips}
\alias{root_2_tips}
\title{Root-to-tips assignment of states to internal nodes}
\usage{
root_2_tips(tree_dat = NA, trait_var = "trait")
}
\arguments{
\item{tree_dat}{dataframe describing phylogeny with internal nodes labelled by tips_to_root.R}
}
\value{
tree_dat with umambiguous moximum parsimony solution to ancestral state reconstruction
}
\description{
Pre-order traversal of phylogeny assigning ancestral states to internal nodes
based on maximum parsimony. Function is designed to be run following a post-order
tips-to-root traversal that determines trait options at each internal node under
a maximum parsimony scheme.
}
\details{
The function requires a data frame with column 'node' and 'parent'.
Such a dataframe is generated using ggtree::fortify
}
