% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_mrca.R
\name{set_mrca}
\alias{set_mrca}
\title{Set Most Recent Common Ancestor}
\usage{
set_mrca(tree_dat = NA, tip_set = NA)
}
\arguments{
\item{tree_dat}{Data frame describing tree structure}

\item{tip_set}{Numeric vector of external/tip nodes}
}
\value{
node number for most recent common ancestor
}
\description{
Given a dataframe describing a phylogenetic tree and a set of tips of
interest, identify the internal node of the tree which represents the
most recent common ancestor (MRCA) of these tips.
}
