% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tips_2_root.R
\name{tips_2_root}
\alias{tips_2_root}
\title{Tips-to-root assignment of states to internal nodes}
\usage{
tips_2_root(
  tree_dat = NA,
  trait_var = "trait",
  prefer_intersect = FALSE,
  resolve_unambiguous = FALSE,
  inferred_column = "trait_infer"
)
}
\arguments{
\item{tree_dat}{Data frame describing tree structure}

\item{trait_var}{Name of column in trait dat with trait to be reconstructed}

\item{prefer_intersect}{Favouring intersect generates more unambiguous internal nodes. Defaults to FALSE}

\item{resolve_unambiguous}{Should a column with inferred trait be returned}

\item{inferred_column}{If resolve_unambigous is TRUE, the name of the column result is returned in}
}
\value{

}
\description{
Perform tips-to-root post-order traversal of phylogenetic tree with labelled discrete
trait in tips. Internal nodes will be labelled parsimoniously. At each internal
node, a list of trait options will be built. There is an option to return
unambiguous node values using this function. An alternative approach is to follow
this function with a root-to-tips traversal (see: root_2_tips.R).
}
