% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_balance_colless.R
\name{tree_balance_colless}
\alias{tree_balance_colless}
\title{Tree Balance using Colless index}
\usage{
tree_balance_colless(tree_dat = NA, branch_length = F, return_df = F)
}
\arguments{
\item{tree_dat}{Data frame describing tree structure}

\item{branch_length}{Logical, implement tree size based version of index}

\item{return_df}{Logical, return a data frame with balance labelled at each internal node}
}
\value{
Colless index or tree data frame with balance labelled for internal nodes
}
\description{
Assess tree balance using the version of the Colless index (Colless, 1982, https://doi.org/10.2307/2413420).
which omits normalisation (Shao & Sokal, 1990, https://doi.org/10.2307/2992186).
Optionally, a version of the index which considers the imbalence between the
size of subtrees descended by each immediate descedant node rather than the
imbalence in the number of tips descended can be implemented using
`branch_length == T`.
}
