% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_balance_sackin.R
\name{tree_balance_sackin}
\alias{tree_balance_sackin}
\title{Tree Balance using Sackin index}
\usage{
tree_balance_sackin(tree_dat = NA, return_df = F)
}
\arguments{
\item{tree_dat}{Data frame describing tree structure}

\item{return_df}{Logical, return a data frame with depth labelled at each external node}
}
\value{
sackin index or tree data frame with depth labelled for external nodes
}
\description{
Tree Balance using Sackin index
}
