% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_node_depth.R
\name{tree_node_depth}
\alias{tree_node_depth}
\title{Node of depth in tree}
\usage{
tree_node_depth(tree_dat = NA)
}
\arguments{
\item{tree_dat}{Data frame describing tree structure}
}
\value{
Input data frame with addition of column denoting depth
}
\description{
For each node calculate its depth, which is the size of the smaller
of the two subtrees descended by either child node, where size is the sum
of branch lengths.
}
