% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_persistence.R
\name{tree_persistence}
\alias{tree_persistence}
\title{Tree persistence}
\usage{
tree_persistence(tree_dat = NA, trait = NA, output_colname = "persistence")
}
\arguments{
\item{tree_dat}{Data frame describing structure of a time-resolved phylogeny}

\item{trait}{Column of tree data on which persistence to be calculated}

\item{output_colname}{Name of persistence column in data frame returned}
}
\value{
tree data frame with column of persistence values
}
\description{
Calculate the phylogenetic persistence of categories within a column of tree
dat. The column should contain character entries and values should have been
generated for internal branches of the phylogeny. For each tip node,
persistence associated with the column categorie is the time back up along
branches of the tree (function requires time-scaled tree).
}
