% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_slice_diversity.R
\name{tree_slice_diversity}
\alias{tree_slice_diversity}
\title{Tree slice diversity}
\usage{
tree_slice_diversity(
  tree_dat = NA,
  tree = NA,
  date_start = NA,
  date_end = NA,
  cond_var = NA,
  cond_val = NA
)
}
\arguments{
\item{tree_dat}{Data frame describing tree structure}

\item{tree}{Phylo class object for tree associated with tree_dat}

\item{date_start}{Lower date limit for temporal slice}

\item{date_end}{Upper date limit for temporal slice}

\item{cond_var}{Column in tree_dat to use for selection}

\item{cond_val}{Value in cond_var column to use for selection}
}
\value{
mean pairwise patristic distance between tips in time slice
}
\description{
Work with a tree and a dataframe describing its structure to calculate the
diversity of tips sampled within a window of time. Diversity is calculated
as the mean of the patristic or phylogenetic distances between tips sampled
during a window in time.
}
