% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_slice_tmrca.R
\name{tree_slice_tmrca}
\alias{tree_slice_tmrca}
\title{Tree slice time of most recent common ancestor}
\usage{
tree_slice_tmrca(
  tree_dat = NA,
  date_start = NA,
  date_end = NA,
  method = "complete",
  cond_var = NA,
  cond_val = NA
)
}
\arguments{
\item{tree_dat}{Data frame describing tree structure}

\item{date_start}{Lower date limit for temporal slice}

\item{date_end}{Upper date limit for temporal slice}

\item{method}{Whether to use the "complete" (default) or "window" method}

\item{cond_var}{Optional column name on which to condition result}

\item{cond_val}{If conditional variable is supplied, requires value to include}
}
\value{
fractional date for tMRCA
}
\description{
Given a dataframe describing a phylogenetic tree, find the time of the most
recent common ancestor of a temporal slice of the tree. Function can find
the time of the most recent common ancestor (tMRCA) of all extant lineages
in the temporal slice (method = "complete") or the tMRCA of tips sampled
during a temporal slice only (method = "window").
}
