% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_slice_unifrac.R
\name{tree_slice_unifrac}
\alias{tree_slice_unifrac}
\title{Tree slice UniFrac calculation}
\usage{
tree_slice_unifrac(
  tree_dat = NA,
  date_start = NA,
  date_end = NA,
  trait_var = "trait",
  per_level = FALSE,
  return_df = FALSE,
  cond_var = NA,
  cond_val = NA
)
}
\arguments{
\item{tree_dat}{Data frame describing tree structure}

\item{date_start}{Lower date limit for temporal slice}

\item{date_end}{Upper date limit for temporal slice}

\item{trait_var}{Name of column with discrete trait}

\item{per_level}{Logical on whether to calculate a level-specific metric rather than the overall UniFrac metric for the trait}

\item{return_df}{Logical - should dataframe with inferred internal nodes be returned}
}
\value{
UniFrac metric
}
\description{
Given a dataframe describing a phylogenetic tree, calculate the UniFrac
metric considering tips within a temporal slice of the tree. UniFrac is a
measure of the phylogenetic signal in a discrete trait.
}
\details{
Nodes are inferred to as being unambiguously associated with
a single trait level or as being ambiguous. The higher the fraction of
unambiguous branches as a proportion of the total branch length in ancestry
of tips within a temporal slice, the greater the phylogenetic signal.
}
