% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunk_lineage.R
\name{trunk_lineage}
\alias{trunk_lineage}
\title{Trunk lineage}
\usage{
trunk_lineage(tree_dat = NA, period = 1)
}
\arguments{
\item{tree_dat}{Data frame describing tree structure}

\item{period}{Period of recent samples whose shared ancestry forms trunk}
}
\value{
set of branches/nodes in trunk lineage
}
\description{
Identify the branches/nodes that form the trunk lineage where the trunk is
defined as the set of branches/nodes that are ancestral to all viruses
sampled within a period (defaults to 1) of the most recent sample.
}
