% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beast_summarise_rates_bssvs.R
\name{beast_summarise_rates_bssvs}
\alias{beast_summarise_rates_bssvs}
\title{Summarise BEAST BSSVS rates}
\usage{
beast_summarise_rates_bssvs(
  rates_dat = NA,
  mean_rate = 1,
  threshold_cond = 0.5
)
}
\arguments{
\item{rates_dat}{Mean overall rate for trait transitions across tree}

\item{threshold}{Value above which to retain conditional means. Defaults to 0.5.}
}
\value{
summary dataframe for transitions and NA matchining dataframe for non-transitions
}
\description{
Summarise info in BEAST generated log file from a run
with discrete traits and a BSSVS network.
}
