% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/does_trait_change.R
\name{does_trait_change}
\alias{does_trait_change}
\title{Does trait change}
\usage{
does_trait_change(
  tree_dat = NA,
  trait_var = "trait",
  logical_column = "trait_change",
  transition_column = NULL,
  separator = "_"
)
}
\arguments{
\item{separator}{}
}
\value{

}
\description{
With a discrete trait labelled in internal nodes
}
