% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_avian_species.R
\name{parse_avian_species}
\alias{parse_avian_species}
\title{Parse avian species}
\usage{
parse_avian_species(host = NA, output = "order")
}
\arguments{
\item{host}{vector of host labels extracted from virus isolate names}

\item{output}{desired output type}
}
\value{
dataframe with tidied species
}
\description{
Tidy up some commonly occurring errors in bird species names in influenza
sequence data bases and match to an avian order when possible
}
