% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_balance_sackin.R
\name{tree_balance_sackin}
\alias{tree_balance_sackin}
\title{Tree Balance using Sackin index}
\usage{
tree_balance_sackin(tree_dat = NA, branch_length = F, return_df = F)
}
\arguments{
\item{tree_dat}{Data frame describing tree structure}

\item{branch_length}{Logical, calculate tree size based version of index}

\item{return_df}{Logical, return a data frame with depth labelled at each external node}
}
\value{
sackin index or tree data frame with depth labelled for external nodes
}
\description{
Function assesses tree balance by calculating the Sackin index which sums
root-to-tip depth across the tree. By default depth is calculated as the
number of edges / branches.
}
