% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_tempest.R
\name{tree_tempest}
\alias{tree_tempest}
\title{Tree temporal estimate}
\usage{
tree_tempest(tree_dat = NA, dates = NA)
}
\arguments{
\item{tree_dat}{Data frame describing tree structure, optionally including a 'date' column}

\item{dates}{Dates can optionally be supplied in a separate data frame}
}
\description{
Investigate the temporal signal in a molecular phylogeny.
Requires a data frame describing tree structure with columns 'branch.length'.
Date info can be supplied as a column 'date' in tree_dat or in a separate data frame.
}
